Thu, May 4, 2023, 7:57 PM -04:00

+++ Week 4: Memory - PSET & Lab/filter-less/helpers.c
@@ -5,7 +5,7 @@
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
 
-   int grayscaleHex;
+   BYTE grayscaleHex;
 
    for (int i = 0; i < height; i++)
    {
